requirements.txt
tk
gitpython